django_webapp/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
Cython==0.29.14
dj-database-url==0.5.0
Django==3.1.7
django-dotenv==1.4.2
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
gensim==3.8.3
gunicorn==20.1.0
idna==2.5
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.5
Pillow==8.1.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
scikit-learn==0.24.1
scipy==1.5.4
sgmllib3k==1.0.0
six==1.15.0
smart-open==4.2.0
soupsieve==2.2.1
sqlparse==0.4.1
textblob==0.15.3
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.59.0
urllib3==1.26.4
whitenoise==5.2.0
